# Conflicts:
#	geles_backend/src/main/resources/import.sql

+++ geles_backend/src/main/java/lt/aigen/geles/controller/CartController.java
@@ -1,5 +1,7 @@
 package lt.aigen.geles.controller;
 
+import lt.aigen.geles.annotations.Authorized;
+import lt.aigen.geles.components.CurrentUser;
 import lt.aigen.geles.models.Cart;
 import lt.aigen.geles.models.Flower;
 import lt.aigen.geles.models.FlowerInCart;
@@ -26,17 +28,18 @@ import java.util.stream.Collectors;
 @RequestMapping("/carts")
 public class CartController {
     CartRepository cartRepository;
-    UserRepository userRepository;
+    CurrentUser currentUser;
     FlowerInCartRepository flowerInCartRepository;
     ModelMapper modelMapper;
 
-    public CartController(CartRepository cartRepository, UserRepository userRepository, FlowerInCartRepository flowerInCartRepository, ModelMapper modelMapper) {
+    public CartController(CartRepository cartRepository, CurrentUser currentUser, FlowerInCartRepository flowerInCartRepository, ModelMapper modelMapper) {
         this.cartRepository = cartRepository;
-        this.userRepository = userRepository;
+        this.currentUser = currentUser;
         this.flowerInCartRepository = flowerInCartRepository;
         this.modelMapper = modelMapper;
     }
 
+    @Authorized
     @GetMapping("/{id}") // /carts/10
     public ResponseEntity<CartDTO> getCart(@PathVariable Long id) {
         var cart = cartRepository.findById(id);
@@ -47,6 +50,7 @@ public class CartController {
         }
     }
 
+    @Authorized
     @PutMapping("/{id}")
     @Transactional
     ResponseEntity<CartDTO> updateCart(@RequestBody @Validated CartDTO cartDTO, @PathVariable Long id) {
@@ -57,6 +61,10 @@ public class CartController {
 
         Cart newCart = oldCart.get();
 
+        if(newCart.getUser().getId() != currentUser.get().getId()){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         for(var f: newCart.getFlowersInCart()){
             flowerInCartRepository.delete(f);
         }
@@ -65,6 +73,7 @@ public class CartController {
         for (var f : cartDTO.getFlowersInCart()) {
             FlowerInCart flowerInCart = convertFromDTO(f);
             flowerInCart.setCart(newCart);
+            flowerInCart.setCartTemplate(null);
             flowerInCart = flowerInCartRepository.save(flowerInCart);
             flowersInCart.add(flowerInCart);
             System.out.println(flowerInCart);
